refactor: move constants to operation directory and update imports
- Move PATH_PATTERNS constants from src/diff/constants.ts to src/operation/constants.ts
- Update operationsSchema.ts to use PATH_PATTERNS.TABLE_BASE instead of hardcoded regex
- Update all import statements across the codebase to use new constants location
- Ensure consistent path pattern validation across the system
