+++ frontend/packages/db-structure/src/diff/columns/__tests__/buildColumnCheckDiffItem.test.ts
@@ -1,7 +1,7 @@
 import type { Operation } from 'fast-json-patch'
 import { describe, expect, it, vi } from 'vitest'
+import { PATH_PATTERNS } from '../../../operation/constants.js'
 import type { Schema } from '../../../schema/index.js'
-import { PATH_PATTERNS } from '../../constants.js'
 import { getChangeStatus } from '../../utils/getChangeStatus.js'
 import { buildColumnCheckDiffItem } from '../buildColumnCheckDiffItem.js'
