+++ frontend/packages/db-structure/src/diff/columns/__tests__/buildColumnNotNullDiffItem.test.ts
@@ -1,7 +1,7 @@
 import type { Operation } from 'fast-json-patch'
 import { beforeEach, describe, expect, it, vi } from 'vitest'
+import { PATH_PATTERNS } from '../../../operation/constants.js'
 import type { Schema } from '../../../schema/index.js'
-import { PATH_PATTERNS } from '../../constants.js'
 import { getChangeStatus } from '../../utils/getChangeStatus.js'
 import { buildColumnNotNullDiffItem } from '../buildColumnNotNullDiffItem.js'
