+++ frontend/packages/db-structure/src/diff/constraints/buildConstraintUpdateConstraintDiffItem.ts
@@ -1,7 +1,7 @@
 import type { Operation } from 'fast-json-patch'
 import { match, P } from 'ts-pattern'
+import { PATH_PATTERNS } from '../../operation/constants.js'
 import type { Schema } from '../../schema/index.js'
-import { PATH_PATTERNS } from '../constants.js'
 import type { ConstraintUpdateConstraintDiffItem } from '../types.js'
 import { getChangeStatus } from '../utils/getChangeStatus.js'
