+++ frontend/packages/db-structure/src/diff/utils/__tests__/getChangeStatus.test.ts
@@ -1,6 +1,6 @@
 import type { Operation } from 'fast-json-patch'
 import { describe, expect, it } from 'vitest'
-import { PATH_PATTERNS } from '../../constants.js'
+import { PATH_PATTERNS } from '../../../operation/constants.js'
 import { getChangeStatus } from '../getChangeStatus.js'
 
 describe('getChangeStatus', () => {
